requirements.txt
psycopg2-binary
requests
python-dotenv
dotenv
schedule
pystray
Pillow
